requirements.txt
# Core Dependencies
boto3==1.34.70
python-dotenv==1.0.1

# Vector Database
chromadb==0.4.24

# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.3

# UI
streamlit==1.32.0

# Utilities
python-multipart==0.0.9
requests==2.31.0